### 2018-10-11 #1014 @arfon
@vmagnin - this pull request should fix the paper compilation: https://github.com/jerryd/gtk-fortran/pull/124
