### 2018-10-11 dab331b
Corrected affiliation in paper.md
