### 2018-10-11 #1014 @vmagnin
> @vmagnin - this pull request should fix the paper compilation: [jerryd/gtk-fortran#124](https://github.com/jerryd/gtk-fortran/pull/124)
@arfon - I have merged your pull request. Effectively, one of the author has no ORCID.
