### 2018-10-12 #1014 @vmagnin
> [ point_right Check article proof page_facing_up point_left ](https://github.com/openjournals/joss-papers/blob/joss.01014/joss.01014/10.21105.joss.01014.pdf)

In the proof, there is a problem with two citations: in the .bib m**a**tcalfe2011 should be replaced by m**e**tcalfe2011. Should I push the correction on the repository of the project ?
Concerning the softwares references, I guess "n.d." means "no date" ?  And when the author is a community, how should the reference appear in the .bib ?
Thanks
