### 2018-10-12 #1014 @Kevin-Mattheus-Moerman
Yes do implement the changes in your own repo. You can call `@whedon generate pdf` yourself here whenever you want to re-render the pdf.

@arfon can you help answer the question about citing a software package when the author is a community, what should be changed in this:
```latex
@misc{plplot,
author = {The Plplot developer community},
title = {PLplot - The ultimate in cross-platform plotting},
howpublished = {\url{http://plplot.sourceforge.net/}}
}
```
to avoid it being rendered as:
`community, T. P. developer. (n.d.). PLplot - the ultimate in cross-platform plotting.
http://plplot.sourceforge.net/.`
