### 2018-10-15 7d52864
JOSS paper: Metcalf reference corrected
