+++ paper.md
@@ -32,9 +32,9 @@ bibliography: paper.bib
 
 # Summary
 
-A Graphical User Interface (GUI) can be useful in a research software to visualize results in real time during a computation, to plot results or to interact with the user, especially if the software is used by other people or distributed outside the laboratory. But despite the venerable age of Fortran, few solutions exist in this language [@markus2012; @metcalfe2012] and are often limited in their functionalities or are not open source, or not cross-platform, etc. For example, you can plot results with PLplot but not create a GUI [@plplot],  DISLIN has a restrictive licence [@dislin], Intel Visual Fortran offers some interesting features for GUI and graphics but is limited to the Microsoft Windows operating system and is commercial software [@visual_fortran]...
+A Graphical User Interface (GUI) can be useful in a research software to visualize results in real time during a computation, to plot results or to interact with the user, especially if the software is used by other people or distributed outside the laboratory. But despite the venerable age of Fortran, few solutions exist in this language [@markus2012; @metcalf2011] and are often limited in their functionalities or are not open source, or not cross-platform, etc. For example, you can plot results with PLplot but not create a GUI [@plplot],  DISLIN has a restrictive licence [@dislin], Intel Visual Fortran offers some interesting features for GUI and graphics but is limited to the Microsoft Windows operating system and is commercial software [@visual_fortran]...
 
-Gtk-fortran [@gtk_fortran] is a GTK+ / Fortran binding using the ISO\_C\_BINDING intrinsic module for interoperability between C and Fortran defined since the Fortran 2003 standard [@markus2012; @metcalfe2012; @fortran]. Although gtk-fortran development is mainly focused on Linux, it can also be built under Windows, Mac OS X... It is licensed under the GNU General Public License version 3. It offers interfaces to around 10000 GTK+ and GLib functions [@gtk], and also to the PLplot library.  A higher-level interface (at a level comparable to the widget routines available in IDL [@idl]) has been developed to offer an easier access to the main GTK+ widgets. Moreover, even though gtk-fortran was mainly developed to visualise computing results and build GUIs, Fortran developers can also access the large set of utility functions offered by the GLib library. Along the wiki documentation of the project, the user can learn to use gtk-fortran thanks to tens of various example programs available in the repository.
+Gtk-fortran [@gtk_fortran] is a GTK+ / Fortran binding using the ISO\_C\_BINDING intrinsic module for interoperability between C and Fortran defined since the Fortran 2003 standard [@markus2012; @metcalf2011; @fortran]. Although gtk-fortran development is mainly focused on Linux, it can also be built under Windows, Mac OS X... It is licensed under the GNU General Public License version 3. It offers interfaces to around 10000 GTK+ and GLib functions [@gtk], and also to the PLplot library.  A higher-level interface (at a level comparable to the widget routines available in IDL [@idl]) has been developed to offer an easier access to the main GTK+ widgets. Moreover, even though gtk-fortran was mainly developed to visualise computing results and build GUIs, Fortran developers can also access the large set of utility functions offered by the GLib library. Along the wiki documentation of the project, the user can learn to use gtk-fortran thanks to tens of various example programs available in the repository.
 
 ![A Fortran program using gtk-fortran to visualize Julia Sets.](paper.png)
 
