+++ paper.bib
@@ -6,9 +6,9 @@
   publisher={Cambridge University Press}
 }
 
-@book{matcalfe2011,
+@book{metcalf2011,
 title={Modern Fortran explained},
-author={Michael Metcalfe, John Reid, Malcolm Cohen},
+author={Michael Metcalf, John Reid, Malcolm Cohen},
 isbn={978-0-19-960142-4},
 year={2011},
 publisher={Oxford University Press}
