### 2018-10-15 55c02b2
JOSS paper: corrected jdelisle affiliation
