### 2018-10-15 dc985d3
JOSS paper: improved reference to Tincher2017 paper
