+++ paper.bib
@@ -86,9 +86,6 @@ title={Progress on real-time, interactive modeling and simulation of nuclear pow
 journal={17th International Topical Meeting on Nuclear Reactor Thermal Hydraulics, NURETH 2017},
 year={2017},
 volume={2017-September},
-note={cited By 0},
-url={https://www.scopus.com/inward/record.uri?eid=2-s2.0-85052390380&partnerID=40&md5=3440dd41e17c62f731839f4c7cf2e253},
-document_type={Conference Paper},
-source={Scopus},
+document_type={Conference Paper}
 }
 
