### 2018-10-15 d1ec585
JOSS paper: added date to DISLIN reference
