### 2018-10-15 #1014 @vmagnin
* Corrected metcalf2011 reference
* Corrected author J. De Lisle affiliation
* improved reference to Tincher2017 paper
* added date to DISLIN reference
