### 2018-10-15 #1014 @vmagnin
@Kevin-Mattheus-Moerman
@arfon

Browsing a lot of JOSS papers, I found one citing the R software using:
```
@Manual{R2018,
	title = {R: A Language and Environment for Statistical Computing},
	author = R Core Team,
	organization = {R Foundation for Statistical Computing},
	address = {Vienna, Austria},
	year = {2018},
	url = {https://www.R-project.org/},
}
```
The result is:
_suppdata is an R (R Core Team 2018) package to provide easy,_
