### 2018-10-15 #1014 @vmagnin
* Tried to improve references
