### 2018-10-15 55a6afa
JOSS paper: improved references
