+++ paper.bib
@@ -15,45 +15,51 @@ publisher={Oxford University Press}
 }
 
 @misc{plplot,
-author = {The Plplot developer community},
+author = {{PLlot team}},
 title = {PLplot - The ultimate in cross-platform plotting},
+year  = {2018},
 howpublished = {\url{http://plplot.sourceforge.net/}}
 }
 
 @misc{dislin,
 author = {Helmut Michels},
 title = {DISLIN},
-year  = {1986--},
+year  = {1986-2018},
 howpublished = {\url{http://www.mps.mpg.de/dislin/}}
 }
 
 @misc{visual_fortran,
 author = {Intel},
 title = {Intel Visual Fortran},
+year  = {2003-2018},
 howpublished = {\url{https://software.intel.com/en-us/fortran-compilers/details#windows}}
 }
 
 @misc{fortran,
-author = {ISO IEC JTC1/SC22/WG5},
+author = {{ISO IEC JTC1/SC22/WG5}},
 title = {The Home of Fortran Standards},
+year  = {2004},
 howpublished = {\url{https://wg5-fortran.org/}}
 }
 
 @misc{gtk_fortran,
-author = {The gtk-fortran developer community},
+author = {{gtk-fortran team}},
 title = {The gtk-fortran home page},
+year  = {2011-2018},
 howpublished = {\url{https://github.com/jerryd/gtk-fortran/wiki}}
 }
 
 @misc{gtk,
-author = {The GTK+ developer community},
+author = {{GTK+ team}},
 title = {The GTK+ Project},
+year  = {1998-2018},
 howpublished = {\url{https://www.gtk.org/}}
 }
 
 @misc{idl,
-author={Harris Geospatial Solutions},
+author={{Harris Geospatial Solutions},
 title={Interactive Data Language},
+year  = {1977-2018},
 howpublished={\url{http://www.harrisgeospatial.com/SoftwareTechnology/IDL.aspx}}
 }
 
@@ -78,6 +84,7 @@ keywords = "Function parser "
 @misc{NetworkMaker,
 author = {Roland Zimm, Isaac Salazar-Ciudad},
 title = {NetworkMaker},
+year  = {2015},
 howpublished = {\url{http://www.biocenter.helsinki.fi/salazar/software.html}}
 }
 
@@ -89,4 +96,3 @@ year={2017},
 volume={2017-September},
 document_type={Conference Paper}
 }
-
