### 2018-10-15 27de67e
JOSS paper: improved references
