+++ paper.bib
@@ -57,7 +57,7 @@ howpublished = {\url{https://www.gtk.org/}}
 }
 
 @misc{idl,
-author={{Harris Geospatial Solutions},
+author={{Harris Geospatial Solutions}},
 title={Interactive Data Language},
 year  = {1977-2018},
 howpublished={\url{http://www.harrisgeospatial.com/SoftwareTechnology/IDL.aspx}}
