### 2018-10-24 #1014 @Kevin-Mattheus-Moerman
@chirayubatra @jlokimlin would you be interested in reviewing this work for the [Journal of Open Source Software (JOSS)](http://joss.theoj.org/)? The [review process](http://joss.theoj.org/about#reviewer_guidelines) focuses on [the software](https://github.com/jerryd/gtk-fortran) and [a short paper](https://github.com/openjournals/joss-papers/blob/joss.01014/joss.01014/10.21105.joss.01014.pdf).
Let me know if you have questions.
