### 2018-11-02 #1014 @Kevin-Mattheus-Moerman
@chirayubatra @jlokimlin @molines @szaghi @komatits @mitotic would you be interested in reviewing [this work](https://github.com/jerryd/gtk-fortran/wiki) for the [Journal of Open Source Software (JOSS)](http://joss.theoj.org/)?

This submission, gtk-fortran, offers _"scientists programming in Fortran a cross-platform library to build Graphical User Interfaces"_.

The JOSS [review process](http://joss.theoj.org/about#reviewer_guidelines) takes place on GitHub and focuses on [the software](https://github.com/jerryd/gtk-fortran) and [a short paper](https://github.com/openjournals/joss-papers/blob/joss.01014/joss.01014/10.21105.joss.01014.pdf).
Let me know if you have questions.
