### 2018-11-02 #1014 @Kevin-Mattheus-Moerman
@johnsamuelwrites (https://orcid.org/0000-0001-8721-7007) thanks for joining as a reviewer! This is a pre-review issue where we gather reviewers. The review process will be initiated later and takes place on a dedicated review issue. In the meantime you can familiarize yourself with the [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#). Let me know if you have questions.
