### 2018-11-02 jerryd/gtk-fortran#125 @johnsamuelwrites
Unable to run the demo example given in the documentation
Following the instructions under the section ['Linux'](https://github.com/jerryd/gtk-fortran/wiki), I encountered errors

<pre>
$ gfortran gtk.f90 ../examples/julia_pixbuf.f90 `pkg-config --cflags --libs gtk+-3.0`
gtk.f90:61:8:

     use g, only: g_signal_connect_data
        1
Fatal Error: Can't open module file ‘g.mod’ for reading at (1): No such file or directory
compilation terminated.
../examples/julia_pixbuf.f90:39:6:

   use gtk, only: gtk_container_add, gtk_drawing_area_new, gtk_events_pending, gtk&
      1
Fatal Error: Can't open module file ‘gtk.mod’ for reading at (1): No such file or directory
compilation terminated.
</pre>

There were no compilation errors and the application was successfully installed.

For details on the review process, refer: https://github.com/openjournals/joss-reviews/issues/1014
