### 2018-11-02 jerryd/gtk-fortran#126 @johnsamuelwrites
Missing information in Documentation
## Unit and functional tests
The documentation does not describe how to run the tests. There is a folder called test with the perl file run_all.pl. But the [main wiki page](https://github.com/jerryd/gtk-fortran/wiki) does not talk about these tests. Also, if perl environment is required, this must also be added in Dependencies section and also in sections related to the different platforms.

## Add links
Under the Download section, it is mentioned

> To download all the project files, you can simply clone the repository using git (see git-basics).

Add a link to git-basics.

One suggestion for the Linux section, use apt instead of apt-get (section Linux)

For details on the review process, refer: openjournals/joss-reviews#1014
