### 2018-11-02 #1014 @vmagnin
> @vmagnin I am unable to access your website URL shown at the top of your project http://gtk+fortran, is it broken?

@Kevin-Mattheus-Moerman
The URL http://gtk+fortran is probably an error in the settings of the project, dating back to the origin of the project... There is no other site than https://github.com/jerryd/gtk-fortran (or its wiki page https://github.com/jerryd/gtk-fortran/wiki).
I will ask @jerryd, who originally created that repository, to delete that URL.

https://github.com/jerryd/gtk-fortran
