### 2018-11-05 #1014 @Kevin-Mattheus-Moerman
@vmagnin currently your paper contains this sentence: `A Graphical User Interface (GUI) can be useful in a research software to visualize results in real time during a computation, to plot results or to interact with the user, especially if the software is used by other people or distributed outside the laboratory.`
Can you please expand your paper a in terms of describing the application of this work to scientific research? Perhaps give some specific examples of how this work is relevant in particular research applications (sketch a picture that helps the reader understand how this work might be important enough to some scientific research that researchers will feel the need to cite it in their paper).
