### 2018-11-16 jerryd/gtk-fortran#125 @vmagnin
Dear John Samuel,

yes, some files were missing in the gfortran compilation command. I have updated the wiki documentation as follows:
> And if you know which gtk-fortran files are needed, you can also simply compile and run each example as follow, from the src/ directory:
> ```bash
> $ gfortran glib-auto.f90 cairo-auto.f90 gdk-auto.f90 gdk-pixbuf-auto.f90 gtk.f90 ../examples/julia_pixbuf.f90 `pkg-config --cflags --libs gtk+-2.0`
> $ ./a.out
> ```
> (please replace 2.0 by 3.0 in the gtk3 branch).
