### 2018-11-16 jerryd/gtk-fortran#126 @vmagnin
Dear John Samuel,

I have added a section in the wiki about testing gtk-fortran:
https://github.com/jerryd/gtk-fortran/wiki#testing-gkt-fortran
Let me know if you find these explications still confused. Perhaps the test.sh script should be moved to the test directory...

Note that there is some perl script because some developers in the project use them, but the users do not need them.

I have added a link in the wiki text toward the "git-basics" page.

I have replaced "apt-get' by "apt" in the wiki.
