### 2018-11-18 jerryd/gtk-fortran#125 @johnsamuelwrites
Thanks for the changes. The example is now working.
