### 2018-11-18 jerryd/gtk-fortran#126 @johnsamuelwrites
Concerning running the test results, I ran the following command
> ./run_all.pl ../build/examples

But I get the following message.
> bash: ./run_all.pl: Permission denied

Do I need any special permissions to run test cases)? Also please add the following lines under the testing section
$ cd test
$  ./run_all.pl ../build/examples
