### 2018-11-20 #1014 @vmagnin
> @vmagnin can you reply to my comment above please. Thanks.

Dear @Kevin-Mattheus-Moerman
sorry for the delay, of course I have read your request and I have begun thinking about the arguments to add in the paper. But I was busy with other things these last weeks.
I will work on it this week.
