### 2018-11-26 e282b07
Modified JOSS paper
