+++ paper.bib
@@ -56,11 +56,29 @@ year  = {1998-2018},
 howpublished = {\url{https://www.gtk.org/}}
 }
 
+@misc{glade,
+author = {{Glade team}},
+title = {Glade - A User Interface Designer},
+year  = {1998-2018},
+howpublished = {\url{https://glade.gnome.org/}}
+}
+
 @misc{idl,
 author={{Harris Geospatial Solutions}},
 title={Interactive Data Language},
 year  = {1977-2018},
 howpublished={\url{http://www.harrisgeospatial.com/SoftwareTechnology/IDL.aspx}}
+}
+
+@article{hunger2008,
+author = {Hunger, Jens and Böhm, Jürgen and Böhlig, Heinz and Hunger, Bernd},
+year = {2008},
+month = {12},
+pages = {745-750},
+title = {Simulation assisted determination of adsorbate structures},
+volume = {174},
+journal = {Studies in Surface Science and Catalysis - STUD SURF SCI CATAL},
+doi = {10.1016/S0167-2991(08)80304-7}
 }
 
 @article{Kanisch2016,
