### 2018-11-26 #1014 @vmagnin
@Kevin-Mattheus-Moerman
I have modified the paper:
* The introduction paragraph has been expanded, and splitted in two. A reference was added.
* Added the following sentences in the (now) third paragraph :
	* _"The gtkf-sketcher tool even makes easier the design of a complex GTK+3 GUI by automatically generating the Fortran code from the XML files created by the user interface designer software Glade [glade]."_
	* _"They can for example implement visual management of tree-like data structures for handling of complex data projects."_

P.S. I hope the "@" mentions will work in this message, because today the github editor does not seem to recognize them...
