### 2018-11-29 #1109 @whedon
[REVIEW]: gtk-fortran: a GTK+ binding to build Graphical User Interfaces in Fortran
**Submitting author:** @vmagnin (<a href="http://orcid.org/0000-0002-9016-9955">Vincent MAGNIN</a>)
**Repository:** <a href="https://github.com/jerryd/gtk-fortran" target ="_blank">https://github.com/jerryd/gtk-fortran</a>
**Version:** v17.10
