### 2018-11-29 #1109 @Kevin-Mattheus-Moerman
@johnsamuelwrites, @devonrex, @sundmanbo this is where the review takes place. There are tick boxes at the top if this issue to guide you through the process. Please let me know if you have any questions. Thanks! :robot: :rocket:
