### 2018-11-30 jerryd/gtk-fortran#126 @johnsamuelwrites
This is working. Now closing the ticket.
