### 2018-12-01 #1109 @sundmanbo
Hi Kevin

I am not sure this is the right way to conatct you
but I am very unhappy about whedon/JOSS sending me a lot of emails about
the reviewing of papers I am not involved in.
How do I stop that?

Bo Sundman

Den tors 29 nov. 2018 kl 10:32 skrev Kevin Mattheus Moerman <
notifications@github.com>:

> @johnsamuelwrites <https://github.com/johnsamuelwrites>, @devonrex
> <https://github.com/devonrex>, @sundmanbo <https://github.com/sundmanbo>
> this is where the review takes place. There are tick boxes at the top if
> this issue to guide you through the process. Please let me know if you have
> any questions. Thanks! 🤖 🚀
>
