### 2018-12-01 #1109 @Kevin-Mattheus-Moerman
@sundmando yes that is annoying. The to of this issue contains instructions to avoid this. I.e. Set yourself as 'Not watching' https://github.com/openjournals/joss-reviews. Let me know if you need more help.
