### 2018-12-01 #1109 @sundmanbo
Yes I need more help.  Where do I find a "No watching" button?

Den lör 1 dec. 2018 kl 11:41 skrev Kevin Mattheus Moerman <
notifications@github.com>:

> @sundmando yes that is annoying. The to of this issue contains
> instructions to avoid this. I.e. Set yourself as 'Not watching'
> https://github.com/openjournals/joss-reviews. Let me know if you need
> more help.
>
