### 2018-12-01 #1109 @arfon
> Yes I need more help.  Where do I find a "No watching" button?

@sundmanbo - if you might want to change your settings in your GitHub profile [here](https://github.com/settings/notifications) to stop emails being sent:

![](https://cloud.githubusercontent.com/assets/4483/20250604/79eb3144-a9de-11e6-9e51-0fe940966689.png)
