### 2018-12-10 #1109 @johnsamuelwrites
@Kevin-Mattheus-Moerman @vmagnin v17.10 is the specified version for the journal. gtk-fortran follows a different versioning scheme, making use of the gtk version as well. For example: https://github.com/jerryd/gtk-fortran/releases shows v17.10.gtk3.22.25 and v17.10.gtk2.24.31. Is it possible to specify these two versions?
