### 2018-12-11 #1109 @vmagnin
> @Kevin-Mattheus-Moerman @vmagnin v17.10 is the specified version for the journal. gtk-fortran follows a different versioning scheme, making use of the gtk version as well. For example: https://github.com/jerryd/gtk-fortran/releases shows v17.10.gtk3.22.25 and v17.10.gtk2.24.31. Is it possible to specify these two versions?

Each version of gtk-fortran is numbered using the Ubuntu version used to build it, for example 17.10: but each release comprises a GTK+ 2 version and a GTK+3 version. And 3.22.25 and 2.24.31 are the precise GTK+ versions used to build gtk-fortran.
