### 2019-01-04 #1109 @sundmanbo
Hi Kevin

I have finished my review and accept the software.  But I had big problems
receiving a lot of messages from reviews of other papers and had to turn
off some obscure feedback buttons inside GITHUB not be be overloaded by
emails.  So maybe I did not finish the review properly.

Best wishes

Bo Sundman

Den fre 4 jan. 2019 kl 11:31 skrev Kevin Mattheus Moerman <
notifications@github.com>:

> @johnsamuelwrites <https://github.com/johnsamuelwrites>, @devonrex
> <https://github.com/devonrex>, @sundmanbo <https://github.com/sundmanbo>
> can you provide an update on the review process? Let me know if I can help
> with anything. Thanks.
>
