### 2019-01-07 #1109 @vmagnin
@Kevin-Mattheus-Moerman

Thank you @johnsamuelwrites for that review. Concerning your three final comments:

1. I clicked on the link 'How to use GTK-Fortran with Code::Blocks IDE' on two different PCs today and it works fine. Perhaps there was a server problem the day you try it ?

2. A quick start tutorial is a good idea and I can work on it during the next months.

3. Fully automating the test run could be a long term project. I guess there is applications to test softwares with GUI, but I have no experience of that kind of software.  If you have some idea on how it could be done, I am interested.

Finally, I can say that such a review process is very interesting because users generally only post bug reports or questions on how to do something with the soft, but there is very rarely comments or critics or even requests concerning the documentation of the soft.
