### 2019-01-08 #1109 @johnsamuelwrites
@Kevin-Mattheus-Moerman @vmagnin Thanks for the response. I verified the link (review comment 1) and it's working now. Regarding review comment 3 on full automation, I am aware that it may be little difficult and may not be the immediate focus of gtk-fortran. As I mentioned in the review, the manual tests are fine, especially considering that this is a GUI building tool and some users may wish to run through all the tests manually.
