### 2019-01-08 #1109 @Kevin-Mattheus-Moerman
@devonrex thanks for letting us know. I'll remove you as reviewer. For future reference the review entails downloading (e.g. the zip file for the repository if you like without using GIT) and testing the software. Then boxes are ticked at the top of this review issue. Editors are also here to help with the process. Thanks for your time.
