### 2019-01-08 #1109 @Kevin-Mattheus-Moerman
@vmagnin since both reviewers are happy with your submission I will recommend that this work is accepted in JOSS.

At this point can you do the following:

- [ ] Create a new release of the reviewed software and archive a version of it e.g. in a service like Zenodo (some find it convenient to use these steps: https://guides.github.com/activities/citable-code/). Once archived please list the DOI of the archived version here.

- [ ] Please inform us on the version tag for the archived release.
