### 2019-01-09 #1109 @vmagnin
Thank you @Kevin-Mattheus-Moerman

- [X] I have created a new release (v17.10.gtk3.22.25_1).
- [ ] Concerning the DOI, I need some time to contact my coauthor @jerryd  who is the owner of the repository. It seems he is the only one who can give Zenodo the access to the repository.
