+++ paper.bib
@@ -46,7 +46,7 @@ howpublished = {\url{https://wg5-fortran.org/}}
 author = {{gtk-fortran team}},
 title = {The gtk-fortran home page},
 year  = {2011-2018},
-howpublished = {\url{https://github.com/jerryd/gtk-fortran/wiki}}
+howpublished = {\url{https://github.com/vmagnin/gtk-fortran/wiki}}
 }
 
 @misc{gtk,
