### 2019-01-19 #1109 @vmagnin
Dear @Kevin-Mattheus-Moerman

I finally succeeded obtaining a DOI:
- I have asked my coauthor jerryd to transfer to me the ownership of the repository. So the new URL of gtk-fortran is https://github.com/vmagnin/gtk-fortran (instead of https://github.com/jerryd/gtk-fortran, which now redirects toward the new URL).
- I have updated that URL in the paper and have regenerated the PDF.
- I have created a new release: v17.10.gtk3.22.25_2
- And have obtained its DOI: 10.5281/zenodo.2544192
