### 2019-02-08 10875c3
CMake will use Doxygen only if available
