### 2019-02-08 b65d1c3
CMake will use Doxygen only if available
