+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,30 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: ''
+labels: Bug
+assignees: vmagnin
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is. Tell which branch of the project is concerned (e.g. gtk3).
+
+**To Reproduce**
+Steps to reproduce the behavior:
+1. 
+2. 
+...
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Screenshots**
+If applicable, add screenshots to help explain your problem.
+
+**Your system:**
+ - OS version: [e.g. Ubuntu 18.04 64 bits]
+ - Compiler version [e.g. gfortran 8.2.0]
+
+**Additional context**
+Add any other context about the problem here.
