### 2019-02-10 #1109 @Kevin-Mattheus-Moerman
@arfon @danielskatz @kyleniemeyer @labarba I recommend we accept this work in JOSS. The new version tag and DOI are:

> * I have created a new release: v17.10.gtk3.22.25_2
> * And have obtained its DOI: 10.5281/zenodo.2544192
