### 2019-02-11 616cf55
remove a tab and update GTK name
