+++ sketcher/gtkf-sketcher.f90
@@ -252,7 +252,7 @@ contains
     !GCC$ ATTRIBUTES DLLEXPORT :: delete_event
     integer(c_int)    :: ret
     type(c_ptr), value :: widget, event, gdata
-	call destroy (widget, gdata)
+    call destroy (widget, gdata)
     ret = FALSE
   end function delete_event
 
@@ -473,7 +473,7 @@ contains
         open(wunit, file=subdir(1:len_trim(subdir))//"_widgets.f90", action='write')
         write(wunit,'(A)')"! "//subdir(1:len_trim(subdir))//" widget module generated by gtkf-sketcher, "//fdate()
         write(wunit,'(A)')"!"
-        write(wunit,'(A)')"! gtkf-sketcher is part of the gtk-fortran GTK+ Fortran Interface Library."
+        write(wunit,'(A)')"! gtkf-sketcher is part of the gtk-fortran GTK Fortran Interface Library."
         write(wunit,'(A)')"!"
         write(wunit,'(A)')"!"
         rewind(60)
@@ -486,7 +486,7 @@ contains
         open(hunit, file=subdir(1:len_trim(subdir))//"_handlers.f90", action='write')
         write(hunit,'(A)')"! "//subdir(1:len_trim(subdir))//" handler module generated by gtkf-sketcher, "//fdate()
         write(hunit,'(A)')"!"
-        write(hunit,'(A)')"! gtkf-sketcher is part of the gtk-fortran GTK+ Fortran Interface Library."
+        write(hunit,'(A)')"! gtkf-sketcher is part of the gtk-fortran GTK Fortran Interface Library."
         write(hunit,'(A)')"!"
         write(hunit,'(A)')"!"
         rewind(60)
@@ -502,7 +502,7 @@ contains
 
       write(50,'(A)')"! "//subdir(1:len_trim(subdir))//" main program generated by gtkf-sketcher, "//fdate()
       write(50,'(A)')"!"
-      write(50,'(A)')"! gtkf-sketcher is part of the gtk-fortran GTK+ Fortran Interface Library."
+      write(50,'(A)')"! gtkf-sketcher is part of the gtk-fortran GTK Fortran Interface Library."
       write(50,'(A)')"!"
       write(50,'(A)')"!"
       rewind(60)
@@ -698,7 +698,7 @@ contains
       write(50,'(A)')"  type(c_ptr) :: error"
       write(50,'(A)')"  error = c_null_ptr"
       write(50,'(A)')""
-      write(50,'(A)')"  ! Initialize the GTK+ Library"
+      write(50,'(A)')"  ! Initialize the GTK Library"
       write(50,'(A)')"  call gtk_init ()"
       write(50,'(A)')""
       write(50,'(A)')"  ! create a new GtkBuilder object"
@@ -745,7 +745,7 @@ contains
       write(50,'(A)')"  ! show the application window"
       write(50,'(A)')"  call gtk_widget_show ("//appwindow(1:len_trim(appwindow))//")"
       write(50,'(A)')""
-      write(50,'(A)')"  ! enter the GTK+ main loop"
+      write(50,'(A)')"  ! enter the GTK main loop"
       write(50,'(A)')"  call gtk_main ()"
       write(50,'(A)')""
       write(50,'(A)')"end program "//subdir(1:len_trim(subdir))
