### 2019-02-12 #1109 @arfon
@vmagnin have you removed the `paper.md` file from your repository? @whedon doesn't seem to be able to find it any longer.
