+++ paper.bib
@@ -1,0 +1,116 @@
+@book{markus2012,
+  title={Modern Fortran in Practice},
+  author={Markus, A.},
+  isbn={9781139510738},
+  year={2012},
+  publisher={Cambridge University Press}
+}
+
+@book{metcalf2011,
+title={Modern Fortran explained},
+author={Michael Metcalf, John Reid, Malcolm Cohen},
+isbn={978-0-19-960142-4},
+year={2011},
+publisher={Oxford University Press}
+}
+
+@misc{plplot,
+author = {{PLlot team}},
+title = {PLplot - The ultimate in cross-platform plotting},
+year  = {2018},
+howpublished = {\url{http://plplot.sourceforge.net/}}
+}
+
+@misc{dislin,
+author = {Helmut Michels},
+title = {DISLIN},
+year  = {1986-2018},
+howpublished = {\url{http://www.mps.mpg.de/dislin/}}
+}
+
+@misc{visual_fortran,
+author = {Intel},
+title = {Intel Visual Fortran},
+year  = {2003-2018},
+howpublished = {\url{https://software.intel.com/en-us/fortran-compilers/details#windows}}
+}
+
+@misc{fortran,
+author = {{ISO IEC JTC1/SC22/WG5}},
+title = {The Home of Fortran Standards},
+year  = {2004},
+howpublished = {\url{https://wg5-fortran.org/}}
+}
+
+@misc{gtk_fortran,
+author = {{gtk-fortran team}},
+title = {The gtk-fortran home page},
+year  = {2011-2018},
+howpublished = {\url{https://github.com/vmagnin/gtk-fortran/wiki}}
+}
+
+@misc{gtk,
+author = {{GTK+ team}},
+title = {The GTK+ Project},
+year  = {1998-2018},
+howpublished = {\url{https://www.gtk.org/}}
+}
+
+@misc{glade,
+author = {{Glade team}},
+title = {Glade - A User Interface Designer},
+year  = {1998-2018},
+howpublished = {\url{https://glade.gnome.org/}}
+}
+
+@misc{idl,
+author={{Harris Geospatial Solutions}},
+title={Interactive Data Language},
+year  = {1977-2018},
+howpublished={\url{http://www.harrisgeospatial.com/SoftwareTechnology/IDL.aspx}}
+}
+
+@article{hunger2008,
+author = {Hunger, Jens and Böhm, Jürgen and Böhlig, Heinz and Hunger, Bernd},
+year = {2008},
+month = {12},
+pages = {745-750},
+title = {Simulation assisted determination of adsorbate structures},
+volume = {174},
+journal = {Studies in Surface Science and Catalysis - STUD SURF SCI CATAL},
+doi = {10.1016/S0167-2991(08)80304-7}
+}
+
+@article{Kanisch2016,
+title = "Generalized evaluation of environmental radioactivity measurements with UncertRadio. Part I: Methods without linear unfolding ",
+journal = "Applied Radiation and Isotopes ",
+volume = "110",
+number = "",
+pages = "28 - 41",
+year = "2016",
+note = "",
+issn = "0969-8043",
+doi = "http://dx.doi.org/10.1016/j.apradiso.2015.12.003",
+author = "Günter Kanisch",
+keywords = "\{ISO\} GUM",
+keywords = "Numerical uncertainty estimation",
+keywords = "Monte Carlo simulation",
+keywords = "ISO 11929 detection limit calculation",
+keywords = "Function parser "
+}
+
+@misc{NetworkMaker,
+author = {Roland Zimm, Isaac Salazar-Ciudad},
+title = {NetworkMaker},
+year  = {2015},
+howpublished = {\url{http://www.biocenter.helsinki.fi/salazar/software.html}}
+}
+
[…]
