### 2019-02-12 #1109 @vmagnin
@arfon The paper.md, paper.bib and paper.png files are still in the master branch of the gtk-fortran repository. But some days ago I have changed the default branch of the repository from master (gtk2) to gtk3. Perhaps @whedon is searching in the default branch.
=> So, I have just copied and pushed the three `paper.*` files in the gtk3 branch. I hope it will solve the problem.

Another reason could be the ownership transfer one month ago from https://github.com/jerryd/gtk-fortran to https://github.com/vmagnin/gtk-fortran, but theoretically everything is automatically redirected from the old URL to the new.
