### 2019-02-12 #1109 @arfon
> @arfon The paper.md, paper.bib and paper.png files are still in the master branch of the gtk-fortran repository. But some days ago I have changed the default branch of the repository from master (gtk2) to gtk3. Perhaps @whedon is searching in the default branch.

Ah yes. @whedon only picks up files in the default branch.
